backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
redis
pyjwt
passlib[bcrypt]
python-dotenv
pydantic
python-multipart
cryptography